Fix the build
Currently we depend on a old version of the SDK that no longer is a the
original location. This will change that so that we now pull the latest
version and run the tests against that. It also upgrades PHP as the
current PHP version for Google App engine is 5.5.